+++ src/index.js
@@ -23,6 +23,7 @@ import BetaFeatures from './containers/BetaFeatures';
 import CrudEndpointUi from './containers/CrudEndpointUi';
 import FlowDebugLogs from './containers/FlowDebugLogs';
 import Flows from './containers/Flows';
+import ForecastDashboards from './containers/ForecastDashboards';
 
 ReactDOM.render(
   <Provider store={store}>
@@ -37,6 +38,7 @@ ReactDOM.render(
               <Navigation />
               <Route path="/interactionMonitoring" component={PrivateRoute(InteractionMonitoring)} />
               <Route path="/agentStateMonitoring" component={PrivateRoute(AgentStateMonitoring)} />
+              <Route path="/forecastDashboards" component={PrivateRoute(ForecastDashboards)} />
               <Route path="/early-access-features" component={PrivateRoute(BetaFeatures)} />
               <Route path="/configuration/:entityName" component={PrivateRoute(CrudEndpointUi)} />
               <Route path="/flowDebugLogs" component={PrivateRoute(FlowDebugLogs)} />