+++ src/redux/epics.js
@@ -27,6 +27,7 @@ import * as customAttributesEpics from './modules/entities/customAttributes/epic
 import * as integrationsEpics from './modules/entities/integrations/epics';
 import * as contactLayoutsEpics from './modules/entities/contactLayouts/epics';
 import * as contactAttributesEpics from './modules/entities/contactAttributes/epics';
+import * as forecastDashboardsEpics from './modules/reporting/forecastDashboards/epics';
 
 export const rootEpic = combineEpics(
   ...Object.values(brandingEpics),
@@ -56,5 +57,6 @@ export const rootEpic = combineEpics(
   ...Object.values(customAttributesEpics),
   ...Object.values(integrationsEpics),
   ...Object.values(contactLayoutsEpics),
-  ...Object.values(contactAttributesEpics)
+  ...Object.values(contactAttributesEpics),
+  ...Object.values(forecastDashboardsEpics)
 );