+++ src/redux/modules/entities/index.js
@@ -185,6 +185,9 @@ const initialState = fromJS({
   standardDashboards: {
     ...defaultEntity
   },
+  forecastDashboards: {
+    ...defaultEntity
+  },
   historicalReportFolders: {
     ...defaultEntity
   },